rename power modifiers as now all ranks
