+++ betterrolls-swade2/scripts/actions/power-modifiers.js
